False positive about $encoding argument for mb_chr(), mb_ord() and mb_scrub().
Fixes issue #1777

These functions were all added in PHP 7.2, so we should not be throwing an
error relating to changes that happened in PHP 5.6.  Prior to this change,
an error would be thrown if the $encoding argument was omitted for these
functions, e.g.:

    The default value of the $encoding parameter for mb_chr() was changed
    from ISO-8859-1 to UTF-8 in PHP 5.6. For cross-version compatibility, the
    $encoding parameter should be explicitly set.
    (PHPCompatibility.ParameterValues.NewIconvMbstringCharsetDefault.NotSet)

I checked all functions covered by this sniff and these were the only ones
that were not present prior to PHP 5.6.

Unit tests have been updated.