PHP 8.4 | :sparkles: New `PHPCompatibility.Syntax.NewClassMemberAccessWithoutParentheses` sniff
>  . new expressions with constructor arguments are now dereferencable, meaning
>    they allow chaining method calls, property accesses, etc. without enclosing
>    the expression in parentheses.
>    RFC: https://wiki.php.net/rfc/new_without_parentheses

This commit adds a new sniff to detect use of this new feature.

The sniff has a close relation with/is similar to the `PHPCompatibility.Syntax.NewClassMemberAccess` sniff and I did consider adding detection of this new PHP 8.4 feature to the pre-existing sniff, though I ultimately decided against this to have more freedom in how to write the sniff and to prevent quite a mess with test changes for the pre-existing sniff.

Includes tests.
Includes documentation.

Refs:
* https://wiki.php.net/rfc/new_without_parentheses
* https://www.php.net/manual/en/migration84.new-features.php#migration84.new-features.core.new-chaining
* php/php-src 13029
* https://github.com/php/php-src/commit/b6b16a1758150cc8992f9a99ac222c5d2998bc1b