PHP 8.3 | Keywords/ForbiddenNames: add support for detecting forbidden names in combination with typed constants
PHP 8.3 introduced typed constants.

While the types themselves should not lead to false positives as the PHP native types are considered "other reserved keywords" and can be used as (class) constant names, so wouldn't be flagged, the sniff as-was would not find the _actual_ name of the constant for typed constants, which could result in false negatives.

Fixed now.

Includes tests.

Ref:
* https://www.php.net/manual/en/reserved.other-reserved-words.php