Various sniffs: prefer using token constants instead of token names
These are remnants of the past when a wider range of PHPCS versions was supported, which meant that some token constants may not exist in all supported PHPCS versions, but this is no longer needed.

Note: this doesn't "fix" this completely. I am aware there are some more places in the codebase where `'type'` is used. These will be reviewed at a later point in time.