Various sniffs: reduce the use of the PHPCSUtils BCTokens class
The PHPCS `Tokens` class provides a number of token arrays for use by sniffs.
The PHPCSUtils `BCTokens` class mirrors that class and provides PHPCS cross-version compatible versions of the same with the latest version from PHPCS being leading in what needs to be supported.

Since both PHPCompatibility, as well as PHPCSUtils are no longer supporting a wide range of PHPCS versions, this has become largely irrelevant, with only some minimal relevance for token arrays which will change in the upcoming PHPCS 4.0 version.

At this time, there are only six token arrays in the `BCTokens` class to which the "PHPCS 4.0 compatility" argument can be applied:
1. `assignmentTokens` - but the cross-version compat is only relevant for libraries which includes sniffs scanning javascript files.
2. `blockOpeners` - but the cross-version compat is only relevant for libraries which includes sniffs scanning javascript files.
3. `functionNameTokens` - provides cross-version compat with the change in how namespaced names are tokenized in PHPCS 4.0.
4. `nameTokens` - new token array in PHPCS 4.0, but was already available in PHPCSUtils via the token arrays in `Collections`.
5. `parenthesisOpeners` - provides cross-version compat with the changes in what tokens will get `parenthesis_*` token array index keys in PHPCS 4.0.
6. `scopeOpeners` - but the cross-version compat is only relevant for libraries which includes sniffs scanning javascript files.

This means that the only two token arrays for which the use of the `BCTokens` variant is relevant for PHPCompatibility are the `functionNameTokens` and the `parenthesisOpeners` token arrays and neither of these are in use by PHPCompatibility.

So this commit does some clean-up, removing the no longer necessary usage of the `BCTokens` class.