FunctionDeclarations/RemovedCallingDestructAfterConstructorExit: start using PHPCSUtils 1.1.0 getDeclaredMethods()
PHPCSUtils 1.1.0 introduces a new `ObjectDeclarations::getDeclaredMethods()` method which can retrieve a list of all methods declared in an OO structure.

This method is highly optimized and will cache its results, which means that if multiple sniffs use the method, the results will be instantaneously returned on subsequent uses.

While this doesn't completely eliminate the need to token walk the OO declaration from within the sniff, it does reduce the token walking significantly, as this now only needs to be done for trait `use` statements if no `destruct()` method was found and the class does not extend anything.
It also allows for bowing out of the token walking as soon as a trait `use` import is found.

This should, again, help with sniff performance.

Includes oem extra tests to ensure all code is covered.