PHP 8.4 | :sparkles: New `PHPCompatibility.Classes.NewAbstractProperties` sniff
As of PHP 8.4, (hooked) properties can be declared as `abstract`.

Notes:
* The "abstract" modifier is not supported in constructor property promotion as that wouldn't make sense.
* The "abstract" modifier is also not supported for interface properties as those are implicitly abstract (but that's not the concern of this sniff).

This commit adds a new sniff to detect abstract properties as supported in PHP 8.4.

For this sniff to function correctly, the PHPCSUtils minimum version needs to be v 1.1.2, which includes support for abstract properties via the `Variables::getMemberProperties()` method.

Includes tests.
Includes documentation.

Refs:
* https://wiki.php.net/rfc/property-hooks#abstract_properties
* php/php-src 13455
* https://github.com/php/php-src/commit/780a8280d23453ebab5df0dbc35ac897c07c1f0d