ResolveHelper::getFQClassNameFromDoubleColonToken(): bug + performance fix - misidentification of class for "self"
If the "self" keyword would be found before a double colon, the utility method would try to get the class name for "self", however, it did not do so correctly.

Previously, the method would try to find the applicable class declaration by walking the tokens to find the first use of the `class` keyword. However, the `self` keyword may be used within an anonymous class, enum, trait etc, in which case, the `class` keyword would never be found, while for enums and interfaces, the name `self` resolves to, can be determined.

On top of that, the `class` keyword found this way may not be the scope in which the `self` keyword is being used, leading to incorrectly resolved names.

This token walking was also highly inefficient and had a negative performance impact.

These problems have now been fixed by:
1. Using the PHPCSUtils `Conditions::getLastCondition()` method to get the applicable OO scope.
    This also effectively removes the token walking.
2. Searching via the `Conditions::getLastCondition()` method for **all** OO scopes and bowing out if the `self` keyword is not found in an OO scope or the OO scope found is an anonymous class or a trait.

Includes tests proving the bug and safeguarding the fix.
All four tests would previously resolve the `self` keyword to the `Unrelated` class.