ResolveHelper::getFQClassNameFromDoubleColonToken(): bug fix - improve comment handling
A comment in an unexpected place, like between the Class name and the double colon, could previously throw off the method and lead to unusable results (class names including the comment, which makes them incomparable for our purposes).

Fixed now by on the one hand allowing for comments while trying to find the start of the name, and on the other hand by using the PHPCSUtils `GetTokensAsString::noEmpties()` method to get the name without stray whitespace or comments.

Includes an additional test proving the bug and safeguarding the fix.