README/docs: update code sample for setting property
PHPCS 3.3.0 introduced a new format for array properties. As support for PHPCS < 3.3.0 has been dropped a while ago, the code sample for how to exclude functions for the `PHPCompatibility.Extensions.RemovedExtensions` sniff can be updated to use the new format.

All the more relevant now as support for the old comma-separated string format will be dropped in PHPCS 4.0.