Interfaces/RemovedSerializable: modernize inline phpcs annotations
PHPCS 3.2.0 introduced the `phpcs:*` annotations to replace the `@codingStandards*` annotations and PHPCS 3.3.0 introduced a way to set array properties via these annotations.

As support for PHPCS < 3.3.0 has been dropped a while ago, the annotations used to test that the public properties are correctly supported can now be updated to use the new format.

All the more relevant now as support for the old annotations will be dropped in PHPCS 4.0.

Notes:
* If a property value was provided inline using the `@codingStandardsChangeSetting` format, a comma-separated string was not automatically parsed to the individual array values.
    For array property values provided using a `phpcs:set` annotation in combination with the `propertyName[]` format, PHPCS will handle the parsing of the comma-separated string to array, so we can now also remove the custom code for this from the sniff.