Numbers/RemovedHexadecimalNumericStrings: allow for PHPCS 4.0/PHP 8.0 namespaced name tokenization
The tokenization of (namespaced) "names" has changed in PHP 8.0 and while this new tokenization was "undone" for PHPCS 3.x, as of PHP_CodeSniffer 4.0.0, the new tokenization is used.

This commit adds handling for the new tokenization of fully qualified function calls to this sniff.

Note: this commit will be more straight forward to review while ignoring whitespace changes.