FunctionUse/ArgumentFunctionsUsage: ignore contents of PHP 7.4+ arrow functions
Arrow functions scopes aren't added to the `conditions` array for tokens within the scope.

This would lead to false positives for this sniff as the "body" of an arrow function should not be regarded as _global_ scope, but was.

Fixed now by skipping over the complete arrow function.

While this means the "call to %s nested in function call" error will also not be thrown, I don't see this as problematic as that error is related to behaviour which is allowed since PHP 5.3 and arrow functions did not exist until PHP 7.4, so the "nested in function call" issue would never be a problem when found in the body of an arrow function anyway.

Includes tests.