FunctionUse/ArgumentFunctionsReportCurrentValue: bug fix - handling of constant in debug*backtrace()
The `DEBUG_BACKTRACE_IGNORE_ARGS` should only be recognized as a reason to ignore the call to the target functions if it refers to the PHP native global constant.

This commit fixes the sniff to not have false negatives if a namespaced constant is used instead of the global one.

Includes tests.