FunctionUse/ArgumentFunctionsReportCurrentValue: bug fix - false negatives for methods and namespaced functions
The sniff contains special casing for when any of the target functions are nested within the PHP native global `array_splice()` or `array_splice()` functions.

However, this special casing did not take namespaced functions or methods mirroring the name of the PHP native functions, into account.

This commit adds more defensive coding to verify we're really dealing with the PHP native global functions.

As the code to check this was already used elsewhere too, it has been moved to a function, so it can be used in both places.

Includes tests.