FunctionUse/ArgumentFunctionsReportCurrentValue: improve handling of array_slice/splice [1]
The offset passed to `array_slice()`/`array_splice()` can be a negative number (offset from the end of the array).

Until now, the sniff did not handle this and would threat all numbers as positives numbers, which could lead to both false positives as well as false negatives.

Fixed now.

Includes tests.