FunctionUse/ArgumentFunctionsReportCurrentValue: handle non-decimal numbers
While probably rare, it is perfectly valid to pass the `$arg_num` parameter for `func_get_arg()` or the `$offset` or `$length` parameters for `array_slice()` and `array_splice()` as non-decimal numbers.

And while the second case (`array_slice()`/`array_splice()`) was fixed in a previous commit, the former (`$arg_num`) was not.

Fixed now.

Includes tests.