FunctionUse/ArgumentFunctionsReportCurrentValue: special case PHP 7.4+ null coalesce assignment
In contrast to all other assignment operators, the null coalesce assignment operator _may_ or _may not_ change the value of a variable.

With that in mind, I believe it is more appropriate to throw a warning when a parameter is being assigned to with the null coalesce operator (in contrast to the error being thrown for other assignment operators).

Includes tests, both for a variety of assignment operators as well as for the null coalesce assignment operator.