FunctionUse/ArgumentFunctionsReportCurrentValue: improve handling of simple control structure conditions
If a function parameter is only used in a control structure condition without further conditions/comparisons, we can be sure it has not been changed.

This commit adds support for detecting this to reduce false positives further.

Notes:
* `foreach()` is special cased as we can ignore use of the variable before the `as`, but should error on use of the variable after the `as`.
* `catch()` is special cased as any variable in catch will always be an assignment, so we should error on that.
* In all other cases, when we're in a control structure condition without other non-empty tokens, we can ignore the use of the variable.

Includes handling of PHP 8.0+ `match()` control structures.

Includes tests.

Loosely related to 796