FunctionUse/ArgumentFunctionsReportCurrentValue: add support for PHP 8.0+ named arguments in function calls
For the `array_*()` functions to was incidentally and partially added in some of the previous commits, though the code still needed to be adjusted to allow for an unexpected parameter order.
For the "target functions", support was not handled so far, so this commit adds support for named parameters.

Verified the parameter name used is in line with the name as per the PHP 8.0 release. PHP itself renamed a lot of parameters in PHP 8.0. As named parameters did not exist before PHP 8.0, the parameter name as per PHP 8.0 (or above) is the only relevant name.

Name verification reference:
* `func_get_arg`: https://3v4l.org/liZU1#veol
* `debug_backtrace`: https://3v4l.org/sFqA7#veol
* `debug_print_backtrace`: https://3v4l.org/LtnLq#veol
* `array_slice`: https://3v4l.org/7hUY7#veol
* `array_splice`: https://3v4l.org/XU9KW#veol

Includes adding the unit tests to include tests using named parameters.

Related to 1239

:point_right: This commit will be more straight-forward to review while ignoring whitespace changes.