Composer: raise the minimum supported PHPCS + PHPCSUtils versions
PHP_CodeSniffer 3.13.3 + 3.13.4 and PHPCSUtils 1.1.2 have been released.
This commit raises the minimum supported PHPCS/Utils versions for improved PHP 8.4 support.

Note: raising to 3.13.4 is not necessary for PHPCompatibility as the fixes includes in that mostly don't affect us.

Includes updating references to the PHPCS version whenever relevant throughout the codebase.

Ref:
* https://github.com/PHPCSStandards/PHPCSUtils/releases/tag/1.1.2
* https://github.com/PHPCSStandards/PHP_CodeSniffer/releases/tag/3.13.3
* https://github.com/PHPCSStandards/PHP_CodeSniffer/releases/tag/3.13.4