ParameterValues/NewExitAsFunctionCall: minor simplification
As of PHPCS 4.0, namespaced names use the tokenization as per PHP 8.0. This also means that we no longer need to check for a `T_NS_SEPARATOR` token for a fully qualified call to exit/die, as that will be tokenized as part of the `T_EXIT` token as of PHPCS 4.0.