Keywords/NewKeywords: split test case file
The test related to `__halt_compiler()` has to be at the end of the file as everything after will not be compiled/tokenized.

In practice, this means that any new tests which need to be added have to be added _above_ the `__halt_compiler()` test and the line number related to the `__halt_compiler()` test needs to be updated every single time.

This makes this test fiddly and prone to getting out of sync, so with this in mind, I'm moving this particular test to its own test case file, which should stabilize the test code for the future.