ParameterValues/NewExitAsFunctionCall: update XML docs
Follow up on 1807 which introduced this sniff.

The initial version of the sniff did not check for FQN exit/die. A check for this was added at a later stage (though before the PR was merged), but the docs were not updated to reflect this new check having been added.

Fixed now.