Classes/RemovedClasses: prevent false positives on imported namespaced classes
Follow up on PRs 1695 and 1700 in which this same change was already made for the `NewClasses`, `InternalInterfaces` and `NewInterfaces` sniffs.

This commit adds rudimentary tracking of the seen import `use` statements in a file to prevent false positives on imported namespaced classes for which the class name overlaps with a removed/deprecated global PHP class name.

Co-authored-by: Sebastian Knott <33318992+rdss-sknott@users.noreply.github.com>