Tests: various minor cleanup
Unless it is to test something specific in how the sniff handling whitespace, sniff test case files should use space indentation. not have trailing whitespace at the end of a line and have a new line at the end of the file.

Note: this is not a complete fix, but a first iteration, though it should get us most of the way there.

:thinking: Might be an idea to add a check for this to the CI ?