PHP 8.5 | RemovedIniDirectives: detect use of intl.error_level (RFC)
> - Intl:
>   . The intl.error_level INI setting has been deprecated.
>     Errors should either be checked manually or exceptions should be enabled
>     by using the intl.use_exceptions INI setting.
>     RFC: https://wiki.php.net/rfc/deprecations_php_8_5#deprecate_intlerror_level_ini_setting

This commit accounts for the deprecation of the ini setting.

In the RFC (and the implementation), the deprecation is conditional on the value:

> we propose deprecating this INI setting, meaning setting it to any value different than `0` will raise an `E_DEPRECATED` diagnostic.

However, for the purpose of PHPCompatibility, I believe checking the value for a call to `ini_set()` is a little over the top. If anyone would report false positives because of this, we can always consider adding such a check later.

Refs:
* https://wiki.php.net/rfc/deprecations_php_8_5#deprecate_intlerror_level_ini_setting
* https://github.com/php/php-src/blob/1570ce9f551f7dff3f51690e5c03b0f5ae0e159f/UPGRADING#L446-L450
* php/php-src 19430
* https://github.com/php/php-src/commit/69f67f990dd3b2eeed9e2687a528b7a00fc2671a

Related to 1849