PHP 8.5 | NewIniDirectives: detect new fatal_error_backtraces ini directive (RFC)
> . Added fatal_error_backtraces to control whether fatal errors should include
>   a backtrace.
>   RFC: https://wiki.php.net/rfc/error_backtraces_v2

This commit accounts for detecting the ini setting.

Note: the PR for the RFC also mentions the following:
> It defaults to `E_FATAL_ERRORS`, which I've also exposed as a userland constant.

However, by the looks of it, this new constant didn't make it into the final implementation (and isn't mentioned in the RFC either), so I've not updated the `NewConstants` sniff.

Refs:
* https://wiki.php.net/rfc/error_backtraces_v2
* https://github.com/php/php-src/blob/34721745833b4c9e69ed22f1b17c4c1c421f7179/UPGRADING#L946-L948
* php/php-src 17056
* https://github.com/php/php-src/commit/0a14ab18d2b3bcf1358aa9f25efc1ad72c18d000

Related to 1849