__sleep/__wakeup: check mutually exclusive methods; add negative tests
There is no need to show a warning when __serialize is implemented in the same class as __sleep, because the first would take precedence, so no behavior change will occur in that context.

Add tests to cover scopes in which these methods are actual magic methods. Currently, these are classes, anonymous classes, and traits. Interfaces were excluded because the value of adding our warning is debatable, since no behavior change will occur.