Add pull request template and improve CONTRIBUTING guide
Based on various things afilina and me discussed in a video call while reviewing one of the new sniffs she pulled.

There are quite some conventions which were not explicit in the CONTRIBUTING guide.
Along the same lines, there are a number of "best practices"/"pro-tips" which can be helpful for people new to contributing to this codebase.

No doubt we'll come across more things to be documented along the way, but this is a first step towards improving the contributor experience.

Co-authored-by: Anna Filina <afilina@gmail.com>