CHANGELOG for the 10.0.0-alpha2 release
References all merged PRs and uses similar grouping as was used in the 10.0.0-alpha1 changelog to keep the changelog human readable.

This release contains one small BC break (`RemovedTypeCasts` sniff error codes). I have prepared the update for the PHPCompatibility 10.0.0 upgrade guide in the wiki and will push that when we release.