Remove LowPHP sniff
This sniff was in place to encourage people to update the PHP version on which they run PHPCompatibility, but the recommended PHP version would currently be the same as the minimum supported PHP version, so the sniff is redundant for now.

If needed, we can always bring it back in the future.