Miscellaneous/RemovedAlternativePHPTags: ASP tags can no longer be detected for sure
What with the version drop of PHP < 7.2, the `asp_tags` ini setting will no longer exist, so we can only detect code which _looks like ASP tags_, but we cannot be 100% sure that they actually _are_ ASP tags.

The sniff already handled that situation and flags the "code under scan" with the `MaybeASPOpenTagFound` when found.
We can now remove the code specific to the sniff detecting ASP tags on PHP < 7.0.